# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zimbabwe Skill Verification Platform API
  version: 1.0.0
  description: API for institution-managed credentials, public verification, blockchain evidence, auditability, and analytics. Examples contain placeholders only. The uploaded PDF is authoritative; a generated PDF is a presentation copy. Credentials become active only after blockchain confirmation, and revoked credentials remain publicly verifiable.
  contact:
    name: API Support
    email: api-support@example.org
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.example.org
    description: Production placeholder
...
paths:
...
  /api/institutions:
    get: { tags: [Institutions], summary: List institutions, operationId: listInstitutions, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer, verifier], parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }, { $ref: '#/components/parameters/SortOrder' }, { name: sortBy, in: query, schema: { type: string, enum: [created_at, name, email, status] } }, { name: status, in: query, schema: { type: string, enum: [active, inactive, 'true', 'false'] } } ], responses: { '200': { $ref: '#/components/responses/Paginated' }, '400': { $ref: '#/components/responses/ValidationError' } } }
    post: { tags: [Institutions], summary: Create institution, operationId: createInstitution, security: [ { BearerAuth: [] } ], x-roles: [super_admin], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/InstitutionInput' } } } }, responses: { '201': { $ref: '#/components/responses/Created' }, '400': { $ref: '#/components/responses/ValidationError' }, '409': { $ref: '#/components/responses/Conflict' } } }
...
  /api/institutions/{id}/blockchain/authorise:
    post: { tags: [Blockchain Administration], summary: Authorise institution wallet on-chain, operationId: authoriseInstitution, security: [ { BearerAuth: [] } ], x-roles: [super_admin], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { $ref: '#/components/responses/Success' }, '422': { $ref: '#/components/responses/ValidationError' }, '503': { $ref: '#/components/responses/Unavailable' } } }
  /api/institutions/{id}/blockchain/deactivate:
    post: { tags: [Blockchain Administration], summary: Deactivate institution wallet on-chain, operationId: deactivateInstitution, security: [ { BearerAuth: [] } ], x-roles: [super_admin], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { $ref: '#/components/responses/Success' }, '503': { $ref: '#/components/responses/Unavailable' } } }
  /api/students:
    get: { tags: [Students], summary: List institution-scoped students, operationId: listStudents, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer, verifier], parameters: [ { $ref: '#/components/parameters/Page' }, { $ref: '#/components/parameters/Limit' }, { $ref: '#/components/parameters/Search' }, { $ref: '#/components/parameters/SortOrder' }, { name: sortBy, in: query, schema: { type: string, enum: [created_at, student_number, full_name, programme] } }, { name: institutionId, in: query, schema: { type: string, format: uuid } }, { name: programme, in: query, schema: { type: string, maxLength: 200 } } ], responses: { '200': { $ref: '#/components/responses/Paginated' }, '403': { $ref: '#/components/responses/AccessDenied' } } }
    post: { tags: [Students], summary: Create student, operationId: createStudent, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudentInput' } } } }, responses: { '201': { $ref: '#/components/responses/Created' }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/AccessDenied' }, '409': { $ref: '#/components/responses/Conflict' } } }
  /api/students/{id}:
    get: { tags: [Students], summary: Get institution-scoped student, operationId: getStudent, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer, verifier], parameters: [ { $ref: '#/components/parameters/Id' } ], responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/AccessDenied' }, '404': { $ref: '#/components/responses/NotFound' } } }
    patch: { tags: [Students], summary: Update institution-scoped student identity, operationId: updateStudent, security: [ { BearerAuth: [] } ], x-roles: [super_admin, institution_admin, issuer], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/StudentUpdateInput' } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '400': { $ref: '#/components/responses/ValidationError' }, '403': { $ref: '#/components/responses/AccessDenied' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' } } }
  /api/students/{id}/institution:
    patch: { tags: [Students], summary: Reassign a credential-free student, description: Super-admin only. Rejected when any credential references the student; credentials are never moved or rewritten., operationId: reassignStudentInstitution, security: [ { BearerAuth: [] } ], x-roles: [super_admin], parameters: [ { $ref: '#/components/parameters/Id' } ], requestBody: { required: true, content: { application/json: { schema: { type: object, required: [institutionId], properties: { institutionId: { type: string, format: uuid } }, additionalProperties: false } } } }, responses: { '200': { $ref: '#/components/responses/Success' }, '403': { $ref: '#/components/responses/AccessDenied' }, '404': { $ref: '#/components/responses/NotFound' }, '409': { $ref: '#/components/responses/Conflict' }, '422': { $ref: '#/components/responses/ValidationError' } } }
...
components:
  securitySchemes:
    BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
    CookieAuth: { type: apiKey, in: cookie, name: verifyzw_session, description: HttpOnly browser session cookie; cookie-authenticated unsafe requests also require X-CSRF-Token. }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    Page: { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    Search: { name: search, in: query, schema: { type: string, maxLength: 200 } }
    SortOrder: { name: sortOrder, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
...
  schemas:
...
    StandardError: { type: object, required: [success, message], properties: { success: { type: boolean, enum: [false] }, code: { type: string, example: ACCESS_DENIED }, message: { type: string, example: You do not have permission to perform this action. }, requestId: { type: string, format: uuid, nullable: true } } }
    ValidationError: { allOf: [ { $ref: '#/components/schemas/StandardError' }, { type: object, properties: { code: { type: string, enum: [VALIDATION_ERROR] }, errors: { type: array, items: { type: object, required: [field, message], properties: { field: { type: string, example: email }, message: { type: string, example: A valid email address is required. } } } } } } ] }
    Pagination: { type: object, required: [page, limit, total, totalPages, hasNextPage, hasPreviousPage], properties: { page: { type: integer }, limit: { type: integer }, total: { type: integer }, totalPages: { type: integer }, hasNextPage: { type: boolean }, hasPreviousPage: { type: boolean } } }
...
    InstitutionInput: { type: object, required: [name, walletAddress, email], properties: { name: { type: string, minLength: 2, maxLength: 200 }, walletAddress: { type: string, pattern: '^0x[a-fA-F0-9]{40}$', example: '<EVM_WALLET_ADDRESS>' }, email: { type: string, format: email }, phone: { type: string, maxLength: 30 } }, additionalProperties: false }
    StudentInput: { type: object, required: [studentNumber, fullName, programme, institutionId], properties: { studentNumber: { type: string, maxLength: 80 }, fullName: { type: string, minLength: 2, maxLength: 150 }, email: { type: string, format: email }, programme: { type: string, minLength: 2, maxLength: 200 }, institutionId: { type: string, format: uuid } }, additionalProperties: false }
    StudentUpdateInput: { type: object, required: [studentNumber, fullName, programme], properties: { studentNumber: { type: string, pattern: '^[A-Za-z0-9][A-Za-z0-9._/-]*$', maxLength: 80 }, fullName: { type: string, minLength: 2, maxLength: 150 }, email: { type: string, format: email, nullable: true }, programme: { type: string, minLength: 2, maxLength: 200 } }, additionalProperties: false }
...
  responses:
    Success: { description: Successful operation, content: { application/json: { schema: { type: object, required: [success], properties: { success: { type: boolean, enum: [true] } }, additionalProperties: true } } } }
    Created: { description: Resource created, content: { application/json: { schema: { type: object, required: [success], properties: { success: { type: boolean, enum: [true] } }, additionalProperties: true } } } }
    Paginated: { description: Paginated result; non-super-admin results are institution scoped where applicable, content: { application/json: { schema: { type: object, required: [success, pagination], properties: { success: { type: boolean, enum: [true] }, pagination: { $ref: '#/components/schemas/Pagination' } }, additionalProperties: true } } } }
...
    ValidationError: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ValidationError' } } } }
...
    AccessDenied: { description: Authenticated actor lacks permission or institution scope, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    NotFound: { description: Resource not found or feature disabled, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
    Conflict: { description: Duplicate or conflicting resource state, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }
...
    Unavailable: { description: Dependency unavailable or operation timed out, content: { application/json: { schema: { $ref: '#/components/schemas/StandardError' } } } }